backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
redis==5.2.1
mcp==1.2.0
openai==1.58.1
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.6
